more lint
SCCS-vsn: sys/kern/kern_resource.c 4.17
SCCS-vsn: sys/kern/sys_generic.c 5.19
SCCS-vsn: sys/ufs/ffs/ffs_alloc.c 2.17
SCCS-vsn: sys/ufs/lfs/lfs_alloc.c 2.17
SCCS-vsn: sys/ufs/ffs/ffs_inode.c 4.29
SCCS-vsn: sys/ufs/ffs/ufs_inode.c 4.29
SCCS-vsn: sys/ufs/lfs/lfs_inode.c 4.29
SCCS-vsn: sys/ufs/ufs/ufs_inode.c 4.29
SCCS-vsn: sys/kern/vfs_lookup.c 4.28
SCCS-vsn: sys/ufs/ffs/ufs_lookup.c 4.28
SCCS-vsn: sys/ufs/ufs/ufs_lookup.c 4.28
SCCS-vsn: sys/kern/vfs_syscalls.c 4.41
SCCS-vsn: sys/ufs/ffs/ffs_vnops.c 4.41
SCCS-vsn: sys/ufs/ffs/ufs_vnops.c 4.41
SCCS-vsn: sys/ufs/lfs/lfs_vnops.c 4.41
SCCS-vsn: sys/ufs/ufs/ufs_vnops.c 4.41

+++ usr/src/sys/kern/kern_resource.c
@@ -1,4 +1,4 @@
-/*	kern_resource.c	4.16	82/10/17	*/
+/*	kern_resource.c	4.17	82/10/19	*/
 
 #include "../h/param.h"
 #include "../h/systm.h"
@@ -230,7 +230,7 @@ getrusage()
 ruadd(ru, ru2)
 	register struct rusage *ru, *ru2;
 {
-	register u_int *ip, *ip2;
+	register long *ip, *ip2;
 	register int i;
 
 	timevaladd(&ru->ru_utime, &ru2->ru_utime);
